refactor: move ZoneBrowserContext logic into FileBrowserContext
